src/requirements.txt
fastapi
uvicorn
PyYAML
itsdangerous
piper-tts
jupyter
openai-whisper
PyJWT
requests